refactor: improved hierarchy
