+++ src/cli.rs
@@ -1,0 +1,1 @@
+pub mod cli_devices_list;
